Refact SCRUM-105: server 컴포넌트 활용을 위해 클라이언트 관련 코드 분리
